compare_module file restructuring
