+++ tests/compare_module/test_nms.py
@@ -1,5 +1,5 @@
 import numpy as np
-from compare_module.compare import mean_bbox, nms
+from compare_module.nms import nms
 
 
 def test_nms():
@@ -9,7 +9,7 @@ def test_nms():
                       [20.0, 5.0, 27.0, 60.0], ])
     scores = np.array([0.9, 0.7, 0.5, 0.1])
     labels = np.array([1, 1, 1, 1])
-    nms_res, _ = nms(boxes, labels, scores, 0.3)
+    nms_res, _ = nms(boxes, labels, scores, 0.3, func=None)
     test_boxes = np.array([[10.0, 15.0, 25.0, 60.0]])
     assert np.array_equal(nms_res, test_boxes)
     boxes = np.array([[10.0, 15.0, 25.0, 60.0],
@@ -18,7 +18,7 @@ def test_nms():
                       [90.0, 80.0, 180.0, 150.0], ])
     scores = np.array([0.9, 0.55, 0.5, 0.6])
     labels = np.array([1, 1, 1, 1])
-    nms_res, _ = nms(boxes, labels, scores, 0.3)
+    nms_res, _ = nms(boxes, labels, scores, 0.3, func=None)
     test_boxes = np.array([[10.0, 15.0, 25.0, 60.0],
                            [90.0, 80.0, 180.0, 150.0]])
     assert np.array_equal(nms_res, test_boxes)
@@ -28,7 +28,7 @@ def test_nms():
                       [90.0, 80.0, 180.0, 150.0], ])
     scores = np.array([0.9, 0.55, 0.5, 0.6])
     labels = np.array([1, 1, 1, 1])
-    nms_res, _ = nms(boxes, labels, scores, 0.3)
+    nms_res, _ = nms(boxes, labels, scores, 0.3, func=None)
     test_boxes = np.array([[10.0, 15.0, 25.0, 60.0],
                            [90.0, 80.0, 180.0, 150.0],
                            [26.0, 65.0, 60.0, 100.0],
@@ -43,7 +43,7 @@ def test_mean_s():
                       [20.0, 5.0, 27.0, 60.0], ])
     scores = np.array([0.9, 0.7, 0.5, 0.1])
     labels = np.array([1, 1, 1, 1])
-    mean_res, _ = nms(boxes, labels, scores, 0.3, func=mean_bbox)
+    mean_res, _ = nms(boxes, labels, scores, 0.3)
     test_boxes = np.array([[10.77, 15.05, 28.5, 59.55]])
     assert np.array_equal(mean_res, test_boxes)
     boxes = np.array([[10.0, 15.0, 25.0, 60.0],
@@ -52,6 +52,6 @@ def test_mean_s():
                       [20.0, 5.0, 27.0, 60.0], ])
     scores = np.array([0.9, 0.7, 0.5, 0.1])
     labels = np.array([1, 5, 1, 1])
-    mean_res, _ = nms(boxes, labels, scores, 0.3, func=mean_bbox)
+    mean_res, _ = nms(boxes, labels, scores, 0.3)
     test_boxes = np.array([[10.67, 16.0, 30.13, 61.67]])
     assert np.array_equal(mean_res, test_boxes)
